refactor: Put the gateway back on the right side of its boundaries
A gateway translates between what the device speaks and what the domain
speaks. Two things in this one were on the wrong side of that line.

ActionEnvelope builds an ActionResult, which is a domain value, so it now
sits beside ActionResult and ActionName rather than in the adapter that
happens to use it. That also puts it where the legacy gateway can reach it
without importing another adapter's helper.

ShellyDeviceGateway read the authentication state cache off the RPC client
with getattr, a lookup the RpcNetworkGateway interface does not promise and
which only worked because the concrete client happens to expose it. The
cache is a dependency, so it is injected, the way LegacyDeviceGateway has
always taken it. A container that forgets to pass it would silently stop
recording which devices ask for credentials, so a wiring test holds it.
